+++ CHANGES.rst
@@ -1,8 +1,20 @@
 History
 =======
 
-1.0b13 (unreleased)
+1.0b12 (unreleased)
 -------------------
+
+- Make sure ``LDAPPrincipals._login_attr`` has a value. This way
+  ``LDAPUsers.id_for_login`` always returns the principal id as stored in the
+  database.
+  [rnix]
+
+- Improve value comparison in ``LDAPAttributesBehavior.__setitem__`` to avoid
+  unicode warnings.
+  [rnix]
+
+- Implement ``invalidate`` on ``node.ext.ldap.ugm._api.Ugm``.
+  [rnix]
 
 - Support for group DNs in ``memberOf`` attribute that are outside of the UGMs configured group.
   [jensens]